refactor: separate session_state_manager into 4 focused modules
Split monolithic session_state_manager.py (490 lines) into:

1. session_models.py (200 lines)
   - SessionStatus enum
   - Turn dataclass
   - SessionState dataclass with state methods
   - _get_utc_now() helper

2. session_manager_base.py (180 lines)
   - SessionManager class with common methods
   - create_session(), get_session(), _get_lock()
   - end_session(), cleanup()
   - get_active_sessions_count(), get_all_sessions_count()

3. session_message_handler.py (80 lines)
   - SessionMessageHandler class for text operations
   - append_message_async() - add message to history
   - increment_utterance_index_async() - increment utterance counter

4. session_audio_handler.py (80 lines)
   - SessionAudioHandler class for audio operations
   - append_audio_chunk() - accumulate audio
   - get_current_audio() - return and clear buffer
   - clear_audio_buffer() - clear audio buffer

5. session_state_manager.py (updated to facade)
   - Now re-exports all symbols for backward compatibility
   - Single import point for all session-related classes

Benefits:
- Clear separation of concerns (models, management, text, audio)
- Easier to test individual handlers
- Reduced coupling between text and audio logic
- Maintains 100% backward compatibility
